Former-commit-id: e9aaad04ada19b3239128d25a77e4139ac69074a

+++ builder/upgrader.py
@@ -3,24 +3,27 @@ import msvcrt
 import os
 import subprocess
 from pathlib import Path
-from typing import Callable, Iterable, Iterator, NamedTuple, Optional, cast
+from typing import Any, Callable, Iterator, NamedTuple, Optional, Sequence, cast
 
 from .color import Stl
 from .env import PythonEnv
 
 
 class Pckg(NamedTuple):
-    req: str
     name: str
     version: str
     url: Optional[str]
 
 
-def format_columns(pckgs: list[Pckg], key_to_str: Callable[[Pckg], str]) -> Iterator[str]:
-    txts = [key_to_str(pckg) for pckg in pckgs]
+def format_columns(objs: Sequence[Any], to_str: Callable[[Any], str], justify: str) -> Iterator[str]:
+    txts = [to_str(obj) for obj in objs]
     len_txt = len(max(txts, key=len))
     for txt in txts:
-        yield f"{txt: <{len_txt + 1}}"
+        yield f"{txt:{justify}{len_txt + 1}}"
+
+
+def format_pckgs_columns(pckgs: Sequence[Pckg], to_str: Callable[[Pckg], str], justify: str) -> Iterator[str]:
+    return format_columns(pckgs, to_str, justify)
 
 
 def line_clear() -> None:
@@ -57,32 +60,37 @@ class Upgrader:
                         print(f"{line}\r", end="")
             line_clear()
 
-    def _to_upgrade(self, req_file: str, eager: bool) -> list[Pckg]:
-        print(Stl.title("Check upgrade for"), Stl.high(req_file))
+    def _to_upgrade(self, *req_files: str, eager: bool) -> list[Pckg]:
+        if req_files:
+            print(
+                Stl.title("Check"),
+                Stl.low("eagerly" if eager else "only needed"),
+                Stl.title("upgrades for"),
+                Stl.low(" and ").join(Stl.high(req_file) for req_file in req_files),
+            )
 
-        report_file = Path(f"{req_file}.json")
-        self._install(
-            "--dry-run",
-            *("-r", req_file),
-            *("--report", report_file.name),
-            *("--upgrade-strategy", "eager" if eager else "only-if-needed"),
-        )
+            report_file = Path("all_requirements.json")
+            self._install(
+                "--dry-run",
+                *("--report", report_file.name),
+                *sum((("-r", req_file) for req_file in req_files), ()),
+                *("--upgrade-strategy", "eager" if eager else "only-if-needed"),
+            )
 
-        if report_file.exists():
-            with report_file.open(mode="r", encoding="utf8") as f:
-                report = json.load(f)
-            report_file.unlink()
+            if report_file.exists():
+                with report_file.open(mode="r", encoding="utf8") as f:
+                    report = json.load(f)
+                report_file.unlink()
 
-            return [
-                Pckg(
-                    req=req_file,
-                    name=data["name"],
-                    version=data["version"],
-                    url=pckg.get("download_info", {}).get("url"),
-                )
-                for pckg in cast(Iterable[dict], report["install"])
-                if (data := pckg.get("metadata"))
-            ]
+                return [
+                    Pckg(
+                        name=data["name"],
+                        version=data["version"],
+                        url=pckg.get("download_info", {}).get("url"),
+                    )
+                    for pckg in cast(Sequence[dict], report["install"])
+                    if (data := pckg.get("metadata"))
+                ]
         return []
 
     @staticmethod
@@ -90,14 +98,12 @@ class Upgrader:
         if (n := len(to_upgrade)) > 0:
             print()
             print(Stl.title("Upgrade:"))
-            for i, (name, version, req) in enumerate(
-                zip(
-                    format_columns(to_upgrade, key_to_str=lambda pckg: Stl.high(pckg.name)),
-                    format_columns(to_upgrade, key_to_str=lambda pckg: Stl.warn(pckg.version)),
-                    format_columns(to_upgrade, key_to_str=lambda pckg: Stl.low(pckg.req)),
-                )
+            for no, name, version in zip(
+                format_columns(range(len(to_upgrade)), lambda i: Stl.title(f" {i + 1}."), ">"),
+                format_pckgs_columns(to_upgrade, lambda pckg: Stl.high(pckg.name), "<"),
+                format_pckgs_columns(to_upgrade, lambda pckg: Stl.warn(pckg.version), "<"),
             ):
-                print(Stl.title(f" {i + 1}."), name, version, Stl.low("from"), req)
+                print(no, name, version)
         return n
 
     def install_for(self, *req_files: str, eager: bool) -> None:
@@ -107,7 +113,7 @@ class Upgrader:
         while msvcrt.kbhit():
             msvcrt.getch()
 
-        to_upgrade = sum((self._to_upgrade(req_file, eager) for req_file in req_files), [])
+        to_upgrade = self._to_upgrade(*req_files, eager=eager)
 
         while True:
             n = self._show_upgrade(to_upgrade)
@@ -133,5 +139,5 @@ class Upgrader:
                     Stl.high(pckg.name),
                     Stl.warn(pckg.version),
                 )
-                self._install(pckg.url or pckg.name)
+                self._install(pckg.url or f"{pckg.name}=={pckg.version}")
                 to_upgrade.remove(pckg)
